+++ src/web-search.ts
@@ -1,9 +1,0 @@
-export {
-	formatPiWebAccessDoctorLines,
-	getPiWebAccessStatus,
-	getPiWebSearchConfigPath,
-	loadPiWebAccessConfig,
-	type PiWebAccessConfig,
-	type PiWebAccessStatus,
-	type PiWebSearchProvider,
-} from "./pi/web-access.js";